+++ src/main/java/com/example/flower_shop/repository/OccasionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OccasionRepository extends JpaRepository<Occasion, Long>{
+public interface OccasionRepository extends JpaRepository<Occasion, Integer>{
 
 }
